io code update
STL file input/output is coming along nicely. Output is the next
milestone in this aspect, but input is taking priority. Will be creating
a base class to represent "feeds" which is inherited by various feed
types as support for them is added. This makes reaper extensible and
more open.

Will be speed testing parsing feed files of various size and seeing how
vectors behave when pushed with thousands of lines. Still designing the
specifics of CFlatFile.

All vector iteration has been replaced with for_each() and lambda
closures.